♻️ refactor: 폴더명 변경 (src/assets/svgs -> src/assets/icons)
