--- src/assets/icons.ts
@@ -1,1 +1,0 @@
-export * from './svgs';
